feat: convert debugfs into loadable kernel module with init/exit, MODULE guards, and wire into kernel boot path (M54)

+++ src/shell/cmds/cmd_cut.c
@@ -62,7 +62,7 @@ void cmd_cut(const char *args) {
             }
             if (cur == field) {
                 size_t flen = strcspn(fp, delim_str);
-                for (size_t k = 0; k < flen; k++) kprintf("%c", (unsigned long)(uint8_t)fp[k]);
+                for (size_t k = 0; k < flen; k++) kprintf("%c", (int)(uint8_t)fp[k]);
             }
             kprintf("\n");
             line = &buf[i + 1];
